+++ extensions/Worker.Extensions.RabbitMQ/release_notes.md
@@ -6,4 +6,4 @@
 
 ### Microsoft.Azure.Functions.Worker.Extensions.RabbitMQ <version>
 
-- <event>
+- <entry>